backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
jinja2==3.1.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==0.6.23
numpy==1.26.4
sqlite-utils==3.37
tqdm==4.66.4
httpx==0.27.0
pdfminer.six==20240706
openai==1.40.0
